시간 자료형 수정

+++ Source/TIMERUN/LevelObjects.h
@@ -39,7 +39,7 @@ public:
 	USkeletalMeshComponent* SkeletalMeshComponent;
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite)
-	int TimeLevel;
+	float TimeLevel;
 
 
 };
